utp-docker/flow/requirements.txt
psycopg2-binary == 2.9.9
promptflow
promptflow-tools